clients: apps router metadata
