+++ clients/apps/web/src/app/maintainer/[organization]/finance/page.tsx
@@ -1,60 +1,19 @@
-'use client'
+import { Metadata, ResolvingMetadata } from 'next'
+import ClientPage from './ClientPage'
 
-import Finance from '@/components/Finance/Finance'
-import { useToast } from '@/components/Toast/use-toast'
-import Head from 'next/head'
-import { useRouter, useSearchParams } from 'next/navigation'
-import {
-  useListAccountsByOrganization,
-  useListPledgesForOrganization,
-  useListRewards,
-} from 'polarkit/hooks'
-import { useEffect } from 'react'
-import { useCurrentOrgAndRepoFromURL } from '../../../../hooks'
+export async function generateMetadata(
+  {
+    params,
+  }: {
+    params: { organization: string }
+  },
+  parent: ResolvingMetadata,
+): Promise<Metadata> {
+  return {
+    title: `${params.organization}`, // " | Polar is added by the template"
+  }
+}
 
 export default function Page() {
-  const router = useRouter()
-  const { org, isLoaded } = useCurrentOrgAndRepoFromURL()
-
-  const params = useSearchParams()
-  const status = params?.get('status')
-
-  const { toast } = useToast()
-
-  useEffect(() => {
-    if (isLoaded && !org) {
-      router.push('/feed')
-      return
-    }
-  }, [isLoaded, org, router])
-
-  useEffect(() => {
-    if (status === 'stripe-connected') {
-      toast({
-        title: 'Stripe setup complete',
-        description: 'Your account is now ready to accept pledges.',
-      })
-    }
-  }, [status, toast])
-
-  const pledges = useListPledgesForOrganization(org?.platform, org?.name)
-  const rewards = useListRewards(org?.id)
-  const accounts = useListAccountsByOrganization(org?.id)
-
-  return (
-    <>
-      <Head>
-        <title>Polar{org ? ` ${org.name}` : ''}</title>
-      </Head>
-      {org && pledges.data?.items && rewards.data?.items && (
-        <Finance
-          pledges={pledges.data.items}
-          rewards={rewards.data.items}
-          org={org}
-          tab="current"
-          accounts={accounts.data?.items || []}
-        />
-      )}
-    </>
-  )
+  return <ClientPage />
 }
